+++ code/Graphics/Mesh.hpp
@@ -31,7 +31,7 @@ public:
 	void init(u32 vertexCount, u32 facesCount);
 	Mesh* addVertex(const Vector3 &vec);
 	Mesh* addNormal(const Vector3 &vec);
-	Mesh* addTexCoord(u32 u, u32 v);
+	Mesh* addTexCoord(f32 u, f32 v);
 	Mesh* addColor(f32 r, f32 g, f32 b, f32 a);
 	Mesh* addFace(u32 v1, u32 v2, u32 v3);
 	void clear();
